+++ src/main/java/com/api/ttoklip/domain/todolist/controller/ToDoListScheduler.java
@@ -1,7 +1,10 @@
-package com.api.ttoklip.domain.todolist.domain;
+package com.api.ttoklip.domain.todolist.controller;
 
 import com.api.ttoklip.domain.member.domain.Member;
 import com.api.ttoklip.domain.member.domain.QMember;
+import com.api.ttoklip.domain.todolist.domain.TodayToDoList;
+import com.api.ttoklip.domain.todolist.domain.TodayToDoListRepository;
+import com.api.ttoklip.domain.todolist.domain.vo.ToDoList;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -15,9 +18,11 @@ public class ToDoListScheduler {
     private final JPAQueryFactory queryFactory;
     private final TodayToDoListRepository todayToDoListRepository;
 
+    private static final String SEOUL = "Asia/Seoul";
+
     @Transactional
     // 매일 자정에 실행 (서울 기준)
-    @Scheduled(cron = "0 0 0 * * *", zone = "Asia/Seoul")
+    @Scheduled(cron = "0 0 0 * * *", zone = SEOUL)
     public void generateTodayToDoList() {
 
         List<Member> allMembers = queryFactory
