+++ src/main/java/com/api/ttoklip/domain/todolist/domain/TodayToDoList.java
@@ -2,6 +2,7 @@ package com.api.ttoklip.domain.todolist.domain;
 
 import com.api.ttoklip.domain.common.base.BaseTimeEntity;
 import com.api.ttoklip.domain.member.domain.Member;
+import com.api.ttoklip.domain.todolist.domain.vo.ToDoList;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
