--- src/main/java/com/api/ttoklip/domain/todolist/domain/TodayToRoListCustomRepository.java
@@ -1,6 +1,0 @@
-package com.api.ttoklip.domain.todolist.domain;
-
-
-public interface TodayToRoListCustomRepository {
-    TodayToDoList findTodayToDoListsByMemberId(final Long memberId);
-}
