+++ src/main/java/com/api/ttoklip/domain/todolist/infrastructure/TodayToDoListJpaRepository.java
@@ -1,0 +1,9 @@
+package com.api.ttoklip.domain.todolist.infrastructure;
+
+import com.api.ttoklip.domain.todolist.domain.TodayToDoList;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TodayToDoListJpaRepository extends JpaRepository<TodayToDoList, Long> {
+
+    TodayToDoList findTodayToDoListsByMemberId(Long memberId);
+}
